+++ src/backends/neon/NeonLayerSupport.cpp
@@ -754,8 +754,8 @@ bool NeonLayerSupport::IsResizeBilinearSupported(const TensorInfo& input,
     descriptor.m_DataLayout = DataLayout::NCHW;
 
     const TensorShape& outputShape = output.GetShape();
-    descriptor.m_TargetHeight = outputShape[2];
-    descriptor.m_TargetWidth  = outputShape[3];
+    descriptor.m_TargetHeight = static_cast<float>(outputShape[2]);
+    descriptor.m_TargetWidth  = static_cast<float>(outputShape[3]);
 
     return IsResizeSupported(input, output, descriptor, reasonIfUnsupported);
 }
